Don't pass db_url anymore since it's not necessary with the connection being supplied.
